+++ NEWS.md
@@ -3,6 +3,7 @@
 - approximate and computationally feasibly 3lvl selection models via the `RoBMA()` and `RoBMA.reg()` functions with the `study_ids` argument when using `algorithm = "ss"`
 - `pooled_effect()` function to compute the pooled effect size from the `RoBMA.reg`, `NoBMA.reg`, and `BiBMA.reg` models
 - `adjusted_effect()` function to compute the adjusted effect size from the `RoBMA.reg`, `NoBMA.reg`, and `BiBMA.reg` models
+- enables `summary_heterogeneity()` for BiBMA models
 
 ### Fixes
 - passing and checks of the `study_ids` and `study_labels` arguments